Interface for StopwatchListOrchestrator
