+++ app/src/main/java/com/gb/stopwatch/stopwatch/StopwatchListOrchestratorImpl.kt
@@ -1,0 +1,50 @@
+package com.gb.stopwatch.stopwatch
+
+import kotlinx.coroutines.*
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+
+class StopwatchListOrchestratorImpl(
+    private val stopwatchStateHolder: StopwatchStateHolder,
+    private val scope: CoroutineScope,
+) : StopwatchListOrchestrator {
+
+    private val mutableTicker = MutableStateFlow("")
+    override val ticker: StateFlow<String> = mutableTicker
+
+    private var job: Job? = null
+
+    override fun start() {
+        if (job == null) startJob()
+        stopwatchStateHolder.start()
+    }
+
+    private fun startJob() {
+        scope.launch {
+            while (isActive) {
+                mutableTicker.value = stopwatchStateHolder.getStringTimeRepresentation()
+                delay(20)
+            }
+        }
+    }
+
+    override fun pause() {
+        stopwatchStateHolder.pause()
+        stopJob()
+    }
+
+    override fun stop() {
+        stopwatchStateHolder.stop()
+        stopJob()
+        clearValue()
+    }
+
+    private fun stopJob() {
+        scope.coroutineContext.cancelChildren()
+        job = null
+    }
+
+    private fun clearValue() {
+        mutableTicker.value = "00:00:000"
+    }
+}
