+++ app/src/main/java/com/gb/stopwatch/viewmodel/MainViewModel.kt
@@ -6,11 +6,8 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.gb.stopwatch.data.TimestampProvider
 import com.gb.stopwatch.data.TimestampProviderImpl
-import com.gb.stopwatch.stopwatch.StopwatchListOrchestrator
-import com.gb.stopwatch.stopwatch.StopwatchStateCalculator
-import com.gb.stopwatch.stopwatch.StopwatchStateHolder
+import com.gb.stopwatch.stopwatch.*
 import com.gb.stopwatch.view.ElapsedTimeCalculator
-import com.gb.stopwatch.stopwatch.TimestampMillisecondsFormatter
 import kotlinx.coroutines.launch
 
 
@@ -21,7 +18,7 @@ internal class MainViewModel(
     private val _textLiveData: MutableLiveData<String> = MutableLiveData()
     val textLiveData: LiveData<String> get() = _textLiveData
 
-    private val stopwatchListOrchestrator = StopwatchListOrchestrator(
+    private val stopwatchListOrchestrator : StopwatchListOrchestrator = StopwatchListOrchestratorImpl(
         StopwatchStateHolder(
             StopwatchStateCalculator(timestampProvider, ElapsedTimeCalculator(timestampProvider)),
             ElapsedTimeCalculator(timestampProvider),
